requirements.txt
tensorflow
opencv-python
matplotlib
numpy
scikit-learn
streamlit
pillow